Name the plugin
I've opened this PR as a suggestion really. Having installed the plugin it is using it's full repo name as the folder which looks a bit messy. `CakephpCurrencyConverter` seems to me like including `Cakephp` when it's inside my app is a bit redundant.

This change will allow it to be installed without it's framework prefix, making the Plugin folder look a bit cleaner.

Thanks